Change gcc optimize settings for 128 KB TinySh
Since commit 19153cc, the PPC software of LWIP-enabled iBob designs has been
built with the "-Os" optimization to minimize code size.  This caused problems
when building with the 128 KB base system, so the optimzation options used are
now "-O2 -fno-reorder-block".  Details below.

Building a LWIP-enabled iBob design (that builds fine with the 64 KB base
system) with the 128 KB base system worked under Windows, but failed under
Linux with error messages like:

ppc405_1/libsrc/lwip_v2_00_a/src/lwip/src/core/mem.c:87: relocation truncated
to fit: R_PPC_EMB_SDA21 against `.sbss'

The exact cause of this error message is unknown, but experimentation has shown
that it is related to the "-Os" (optimize for size) option to gcc.

Building the LWIP-enabled iBob design with the 64 KB base system (which uses
"-Os") resulted in an executable.elf file with text+data sections of 63148+2344
= 65492 bytes, just under the max size of 65532 (4 bytes left over for boot
vector).  Building the same design with the 128 KB base system works so long as
"-Os" is not used.  I think maybe "-Os" performs some size optimization that
requires data in the .sbss region to be within 64 KB of the code accessing it.
With the 64 KB base system, the code is stored starting 64 KB down from the top
of memory space and the .sbss section is stored at the bottom of memory space.
With the 128 KB base system, the code is stored starting 128 KB down from the
top of memory space.

Here are the resulting test+data sizes of the same design built using the 128
KB base system (except -Os, which was built using the 64 KB base system) and
various compiler optimzation settings:

-O0 : 76756 + 2344 = 79100
-O1 : 65304 + 2344 = 67648
-O2 : 66072 + 2344 = 68416
-O3 : 71752 + 2344 = 74096
-Os : 63148 + 2344 = 65492 <-- Smallest

The -Os optimization level is equivalent to -O2 minus options that increase
code size plus, as the GCC manual vaguely states, "further optimizations
designed to reduce code size".  Of all the -O2 optimzations disabled by -Os, I
found that -fno-reorder-blocks was the only one that affected code size for the
iBob TinySh software.  Here are the resulting test sizes of the same design
built using the 128 KB base system and various compiler optimization settings
plus -fno-reorder-blocks:

-O0 -fno-reorder-blocks : 76756 + 2344 = 79100
-O1 -fno-reorder-blocks : 65304 + 2344 = 67648
-O2 -fno-reorder-blocks : 64816 + 2344 = 67160 <-- Smallest
-O3 -fno-reorder-blocks : 70496 + 2344 = 72840

That is why the optimization settings for LWIP-enabled iBob designs using the
128 KB base system were changed from "-Os" to "-O2 -fno-reorder-blocks".